usage-modeling/requirements.txt
pyyaml==6.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
